illusion_analyzer.py
import os
import pickle
import torch
from PIL import Image
import torchvision.transforms as transforms
from pathlib import Path
import json
from typing import List, Dict, Any
import logging
from datetime import datetime

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler(f'llm_analysis_{datetime.now().strftime("%Y%m%d_%H%M%S")}.log'),
        logging.StreamHandler()
    ]
)

class IllusionAnalyzer:
    def __init__(self, model_name: str = "gpt-4-vision-preview"):
        """
        Initialize the illusion analyzer with an LLM model
        
        Args:
            model_name: Name of the LLM model to use
        """
        self.model_name = model_name
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
        ])
        
    def load_illusion_data(self, illusion_dir: str) -> Dict[str, Any]:
        """
        Load illusion images and metadata
        
        Args:
            illusion_dir: Directory containing the illusion data
            
        Returns:
            Dictionary containing images and metadata
        """
        illusion_path = Path(illusion_dir)
        
        # Load metadata
        with open(illusion_path / 'metadata.pkl', 'rb') as f:
            metadata = pickle.load(f)
            
        # Load images
        images = {}
        for img_path in illusion_path.glob('**/sample_*.png'):
            if 'views' not in str(img_path):  # Skip the views image
                size = img_path.stem.split('_')[1]
                images[size] = Image.open(img_path)
                
        return {
            'metadata': metadata,
            'images': images
        }
    
    def analyze_illusion(self, illusion_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Analyze an illusion using the LLM
        
        Args:
            illusion_data: Dictionary containing images and metadata
            
        Returns:
            Dictionary containing analysis results
        """
        metadata = illusion_data['metadata']
        images = illusion_data['images']
        
        # Get original prompts and views
        original_prompts = metadata['args'].prompts
        views = metadata['args'].views
        style = metadata['args'].style
        
        analysis_results = {
            'original_prompts': original_prompts,
            'views': views,
            'style': style,
            'llm_analysis': {},
            'reasoning_process': []
        }
        
        # Analyze each image size
        for size, image in images.items():
            logging.info(f"Analyzing {size}x{size} image")
            
            # Prepare image for LLM
            image_tensor = self.transform(image)
            
            # Get LLM analysis
            analysis = self._get_llm_analysis(image_tensor, views)
            analysis_results['llm_analysis'][size] = analysis
            
            # Track reasoning process
            analysis_results['reasoning_process'].append({
                'size': size,
                'analysis': analysis,
                'timestamp': datetime.now().isoformat()
            })
            
        return analysis_results
    
    def _get_llm_analysis(self, image: torch.Tensor, views: List[str]) -> Dict[str, Any]:
        """
        Get analysis from LLM for a single image
        
        Args:
            image: Image tensor
            views: List of view transformations
            
        Returns:
            Dictionary containing LLM analysis
        """
        # TODO: Implement actual LLM call here
        # This is a placeholder for the actual LLM implementation
        return {
            'description': "Placeholder for LLM description",
            'identified_views': views,
            'confidence_scores': {},
            'reasoning': "Placeholder for LLM reasoning"
        }
    
    def evaluate_analysis(self, analysis_results: Dict[str, Any]) -> Dict[str, Any]:
        """
        Evaluate how well the LLM analysis matches the original prompts
        
        Args:
            analysis_results: Dictionary containing analysis results
            
        Returns:
            Dictionary containing evaluation metrics
        """
        original_prompts = analysis_results['original_prompts']
        llm_analyses = analysis_results['llm_analysis']
        
        evaluation = {
            'prompt_matching_scores': {},
            'view_identification_accuracy': {},
            'overall_confidence': {}
        }
        
        # TODO: Implement actual evaluation metrics
        # This would include:
        # 1. Semantic similarity between LLM descriptions and original prompts
        # 2. Accuracy of view identification
        # 3. Confidence scores for each analysis
        
        return evaluation
    
    def save_analysis(self, analysis_results: Dict[str, Any], output_dir: str):
        """
        Save analysis results to file
        
        Args:
            analysis_results: Dictionary containing analysis results
            output_dir: Directory to save results
        """
        output_path = Path(output_dir)
        output_path.mkdir(exist_ok=True, parents=True)
        
        # Save analysis results
        with open(output_path / 'analysis_results.json', 'w') as f:
            json.dump(analysis_results, f, indent=2)
            
        # Save evaluation metrics
        evaluation = self.evaluate_analysis(analysis_results)
        with open(output_path / 'evaluation_metrics.json', 'w') as f:
            json.dump(evaluation, f, indent=2)

def main():
    # Example usage
    analyzer = IllusionAnalyzer()
    
    # Load and analyze an illusion
    illusion_data = analyzer.load_illusion_data('results/rotate_cw.village.horse')
    analysis_results = analyzer.analyze_illusion(illusion_data)
    
    # Save results
    analyzer.save_analysis(analysis_results, 'analysis_output')

if __name__ == "__main__":
    main() 