requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
transformers>=4.30.0
openai==0.28.1
sentence-transformers>=2.2.0
python-dotenv>=0.19.0
lime>=0.2.0.1
matplotlib>=3.5.0
scikit-image>=0.19.0 